feat: subcommand folder was renamed to controllers
