+++ src/controllers/cmd_setup.rs
