+++ src/controllers/cmd_stress.rs
